refactor: flatten web feature modules and query layers
