+++ apps/web/src/app/admin/page.tsx
@@ -1,3 +1,3 @@
-import { AdminPage } from "@/features/admin/pages/admin-page";
+import { AdminPage } from "@/features/admin/admin-page";
 
 export default AdminPage;
