+++ apps/web/src/features/admin/admin-page.tsx
@@ -1,8 +1,6 @@
 import { headers } from "next/headers";
-
-import { requireCurrentUser } from "@/features/auth/lib/server";
-
-import { AdminDashboard } from "../components/admin-dashboard";
+import { requireCurrentUser } from "@/features/auth/auth.server";
+import { AdminDashboard } from "./components/admin-dashboard";
 
 export const AdminPage = async () => {
   await requireCurrentUser(await headers());
