+++ apps/web/src/features/anime/anime.hooks.ts
@@ -1,16 +1,16 @@
 "use client";
 
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
+import type { UpsertAnimeBody } from "@anilog/contracts";
 import { useEffect, useState } from "react";
 import { toast } from "sonner";
-import type { UpsertAnimeBody } from "@anilog/contracts";
-import { animeKeys } from "@/features/anime/api/anime.keys";
 import {
+  animeMutations,
   animeQueries,
   type AnimeArchiveSearchInput,
   type AnimeSearchInput,
-} from "@/features/anime/api/anime.query";
-import { animeMutations } from "@/features/anime/api/anime.mutation";
+} from "./anime.api";
+import { animeKeys } from "./anime.keys";
 import { getApiErrorMessage } from "@/lib/eden-fetch";
 
 export function useTrendingAnime() {
