--- apps/web/src/features/anime/api/anime.mutation.ts
@@ -1,8 +1,0 @@
-import type { UpsertAnimeBody } from "@anilog/contracts";
-import { animeClient } from "./anime.client";
-
-export const animeMutations = {
-  upsertAnime: () => ({
-    mutationFn: ({ body }: { body: UpsertAnimeBody }) => animeClient.upsert(body),
-  }),
-};
