+++ apps/web/src/features/anime/components/anime-card.tsx
@@ -4,8 +4,7 @@ import Image from "next/image";
 
 import { cn } from "@/lib/utils";
 import { Button } from "@/components/ui/button";
-import type { LibraryStatus } from "@/features/library/api/library.query";
-import type { LogAnimeData } from "@/features/library/api/library.mutation";
+import type { LibraryStatus, LogAnimeData } from "@/features/library/library.api";
 
 interface AnimeCardProps {
   anime: Pick<LogAnimeData["anime"], "id" | "title" | "imageUrl" | "year" | "episodes"> & {
