+++ apps/web/src/features/library/library.server.ts
@@ -1,7 +1,7 @@
 import "server-only";
 
 import type { QueryClient } from "@tanstack/react-query";
-import { libraryQueries } from "../api/library.query";
+import { libraryQueries } from "./library.api";
 
 export async function prefetchMyLibrary(queryClient: QueryClient) {
   await queryClient.prefetchQuery(libraryQueries.myLibrary());
