--- apps/web/src/features/users/hooks.ts
@@ -1,126 +1,0 @@
-import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
-import { toast } from "sonner";
-import { userMutations } from "@/features/users/api/user.mutation";
-import { userKeys } from "@/features/users/api/user.keys";
-import { userQueries, type UserWithProfile } from "@/features/users/api/user.query";
-import { getApiErrorMessage } from "@/lib/eden-fetch";
-
-export const useSearchUsers = (query: string) => {
-  return useQuery({
-    ...userQueries.search({ query: { q: query } }),
-    enabled: query.length >= 3,
-  });
-};
-
-export const useUserProfile = (userId: string) => {
-  return useQuery({
-    ...userQueries.profile({ params: { id: userId } }),
-    enabled: !!userId,
-  });
-};
-
-export const useUserByUsername = (username: string) => {
-  return useQuery({
-    ...userQueries.byUsername({ params: { username } }),
-    enabled: !!username,
-  });
-};
-
-export const useUserLists = (userId: string) => {
-  return useQuery({
-    ...userQueries.publicLibrary({ params: { id: userId } }),
-    enabled: !!userId,
-  });
-};
-
-export const useIsFollowing = (userId: string, options?: { enabled?: boolean }) => {
-  return useQuery({
-    ...userQueries.isFollowing({ params: { id: userId } }),
-    enabled: !!userId && (options?.enabled ?? true),
-  });
-};
-
-export const useFollowUser = () => {
-  const queryClient = useQueryClient();
-
-  return useMutation({
-    ...userMutations.follow(),
-    onSuccess: (_, { params }) => {
-      toast.success("Successfully followed user!");
-      queryClient.setQueryData(userKeys.isFollowing(params.id), { isFollowing: true });
-      queryClient.setQueryData<UserWithProfile>(
-        userKeys.profile(params.id),
-        (current: UserWithProfile | undefined) =>
-          current ? { ...current, followerCount: current.followerCount + 1 } : current,
-      );
-
-      queryClient.invalidateQueries({ queryKey: userKeys.searchRoot() });
-      queryClient.invalidateQueries({ queryKey: userKeys.profileRoot() });
-      queryClient.invalidateQueries({ queryKey: userKeys.byUsernameRoot() });
-      queryClient.invalidateQueries({ queryKey: userKeys.meProfile() });
-      queryClient.invalidateQueries({ queryKey: userKeys.following() });
-    },
-    onError: (error) => {
-      toast.error(getApiErrorMessage(error));
-    },
-  });
-};
-
-export const useUnfollowUser = () => {
-  const queryClient = useQueryClient();
-
-  return useMutation({
-    ...userMutations.unfollow(),
-    onSuccess: (_, { params }) => {
-      toast.success("Successfully unfollowed user");
-      queryClient.setQueryData(userKeys.isFollowing(params.id), { isFollowing: false });
-      queryClient.setQueryData<UserWithProfile>(
-        userKeys.profile(params.id),
-        (current: UserWithProfile | undefined) =>
-          current ? { ...current, followerCount: Math.max(0, current.followerCount - 1) } : current,
-      );
-
-      queryClient.invalidateQueries({ queryKey: userKeys.searchRoot() });
-      queryClient.invalidateQueries({ queryKey: userKeys.profileRoot() });
-      queryClient.invalidateQueries({ queryKey: userKeys.byUsernameRoot() });
-      queryClient.invalidateQueries({ queryKey: userKeys.meProfile() });
-      queryClient.invalidateQueries({ queryKey: userKeys.following() });
-    },
-    onError: (error) => {
-      toast.error(getApiErrorMessage(error));
-    },
-  });
-};
-
-export const useMyFollowing = () => {
-  return useQuery(userQueries.myFollowing());
-};
-
-export const useMyProfile = () => {
-  return useQuery(userQueries.myProfile());
-};
-
-export const useMyAdminStatus = (options?: { enabled?: boolean }) => {
-  return useQuery({
-    ...userQueries.myAdminStatus(),
-    enabled: options?.enabled ?? true,
-  });
-};
-
-export const useUpdateMyProfile = () => {
-  const queryClient = useQueryClient();
-
-  return useMutation({
-    ...userMutations.updateMyProfile(),
-    onSuccess: () => {
-      toast.success("Profile updated successfully!");
-      queryClient.invalidateQueries({ queryKey: userKeys.meProfile() });
-      queryClient.invalidateQueries({ queryKey: userKeys.profileRoot() });
-      queryClient.invalidateQueries({ queryKey: userKeys.byUsernameRoot() });
-      queryClient.invalidateQueries({ queryKey: userKeys.searchRoot() });
-    },
-    onError: (error) => {
-      toast.error(getApiErrorMessage(error));
-    },
-  });
-};
